% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentmle1.R
\name{gentmle1}
\alias{gentmle1}
\title{gentmle}
\usage{
gentmle1(initdata, estimate_fun, update_fun, max_iter = 100, ...)
}
\arguments{
\item{estimate_fun}{Function for estimation step}

\item{update_fun, }{Function for update step}

\item{max_iter, }{Maximum number of iteration steps}

\item{..., }{Extra arguments that can be passed to update_fun and estimate_fun}
}
\description{
General TMLE function that takes care of the bookkeeping of estimation and update steps.
}
