% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentmle_alt.R
\name{gentmle_alt}
\alias{gentmle_alt}
\title{gentmle}
\usage{
gentmle_alt(initdata, estimate_fun, update_fun, max_iter = 100, N = NULL,
  simultaneous.inference = TRUE, kernel = NULL, blip = NULL, h = NULL,
  ...)
}
\arguments{
\item{estimate_fun}{Function for estimation step}

\item{update_fun, }{Function for update step}

\item{max_iter, }{Maximum number of iteration steps}

\item{N, }{controls order criterion, currently not used}

\item{simultaneous.inference, }{passed to the}

\item{..., }{Extra arguments that can be passed to update_fun and estimate_fun}
}
\description{
General TMLE function that takes care of the bookkeeping of estimation and update steps.
}
