% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentmle_alt2.R
\name{gentmle_alt2}
\alias{gentmle_alt2}
\title{gentmle}
\usage{
gentmle_alt2(initdata, estimate_fun, update_fun, max_iter = 100, N = NULL,
  t, h, k, simultaneous.inference = TRUE, ...)
}
\arguments{
\item{estimate_fun}{Function for estimation step}

\item{update_fun, }{Function for update step}

\item{max_iter, }{Maximum number of iteration steps}

\item{..., }{Extra arguments that can be passed to update_fun and estimate_fun}
}
\description{
General TMLE function that takes care of the bookkeeping of estimation and update steps.
}
